+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.3.0-rc.32",
-  genVersion: "2.808.1",
-  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.32 2.808.1 2.0 @orq-ai/node",
+  sdkVersion: "4.3.0-rc.33",
+  genVersion: "2.809.2",
+  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.33 2.809.2 2.0 @orq-ai/node",
 } as const;