+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -76,7 +76,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kgqb2h86xp64sfajw61d6ax6"),
+  _id: z.string().default("reasoning_01kgrpz9h800cghywnzh4j07ze"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),