+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -280,7 +280,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kgqb2hn3g4azwncmp8e37d10"),
+  _id: z.string().default("reasoning_01kgrpz9y7jpzq65hxn2w44dt4"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -366,7 +366,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kgqb2hmy67g7jkq4jdbfxgc3"),
+  _id: z.string().default("reasoning_01kgrpz9y16247wd84yt8cajax"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),