+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-05T16:44:41.846Z",
+    "2026-02-06T05:31:53.389Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {