+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -180,7 +180,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kgqb2hkz5498qzx20vxj1d5h"),
+  _id: z.string().default("conv_01kgrpz9x46r99khg84t6ektw5"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),