+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -176,7 +176,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KGQB2J5E1236A3E7500R3NS8"),
+  _id: z.string().default("01KGRPZAETZ325F59S0K9CTB6P"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,