+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -226,7 +226,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kgqb2hmm85amk9qsmf51j95b"),
+  _id: z.string().default("conv_01kgrpz9xend7qj3yrvd2pa90m"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),